pyproject.toml
[project]
name = "common-project"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy<2",
    "pandas>=2.2.3",
    "scikit-surprise>=1.1.4",
    "streamlit>=1.45.0",
]
